requirements.txt
openai
playwright
pydantic
python-dotenv
rich

